refactor: BaseStatusModalコンポーネントを作成し、モーダルの共通化を行う
